requirements.txt
aiogram==3.0.0
mysql-connector-python
matplotlib